Working non-TLS type secrets.
When working with OpenSearch I need to create an api certificate, a
certificate for a transport layer and client certificate for
administration of the cluster. These are all kept in the same secret so
they can be mounted in a directory next to each other.

Further more, the keys need be in PKCS#8 format.

Added an annotation for a colon separated list of key pairs in secret.
The pairs are slash delimited. Shouldn't be the case that someone needs
a slash or a colon in their key name so I'm not going to add escaping
for the delimiters. The pair can contain a third part which is the key
type either `pkcs8` or `pem`, but I'm not sure if `pem` is the right
name for the default format. If no pair string is provided it assumed to
be `"tls.crt/tls.key/pem"`.